apps/api/pyproject.toml
[project]
name = "master-clash-api"
version = "0.1.0"
description = "Master Clash API Server - AI-powered video production toolkit"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "hrhrng", email = "hrhrng@foxmail.com" }
]
keywords = ["ai", "video-production", "langchain", "langgraph", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

# Core dependencies
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # AI/LLM
    "dotenv>=0.9.9",
    "google-genai>=1.52.0",
    "langchain>=1.0.3",
    "langchain-google-genai>=3.2.0",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "langgraph-checkpoint>=2.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langsmith>=0.4.47",
    "openai>=0.27.10",
    # Data & Utils
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "tabulate>=0.9.0",
    # Database
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    # Cloud Storage
    "boto3>=1.35.0",
    "requests>=2.32.5",
    "deepagents>=0.2.8",
    "clash-types",
    # Video Analysis
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scenedetect[opencv]>=0.6.4",
    "scikit-image>=0.21.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "langchain-google-vertexai",
    "loro>=1.10.3",
    "aioboto3>=15.5.0",
    "google-cloud-storage>=3.7.0",
    "gcloud-aio-storage>=9.6.1",
]

# Optional dependency groups
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
notebook = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.0",
]
postgres = [
    "psycopg[binary]>=3.1",
]

# Project URLs
[project.urls]
Homepage = "https://github.com/yourusername/master-clash"
Repository = "https://github.com/yourusername/master-clash"
Issues = "https://github.com/yourusername/master-clash/issues"

# Build system
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/master_clash"]

# Ruff configuration - Modern Python linter and formatter
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]
extend-exclude = [
    "**/.history/**",
    "../../apps/web/script/workflow.ipynb",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=master_clash",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# MyPy configuration - Static type checker
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true when adding type hints
ignore_missing_imports = true
strict_optional = true

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "**/__init__.py",
    "**/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
    "types-requests",
]
